pyproject.toml
[tool.poetry]
name = "iaas-jwt"
version = "0.1.0"
description = ""
authors = ["Reinaldo Saraiva do Carmo <reinaldo.carmo@luizalabs.com>"]
readme = "README.md"
packages = [{include = "iaas_jwt"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "7.2.1"
pytest-cov = "4.0.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.10.4"
pyjwt = {extras = ["crypto"], version = "^2.6.0"}
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
ipython = "^8.12.0"
ipykernel = "^6.22.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.6"
mkdocstrings = "^0.21.2"
mkdocstrings-python = "^0.9.0"
griffe = "^0.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"
[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue . && isort ."
docs = "mkdocs serve" 
pre_test = "task lint"
test = "pytest -s -x --cov-report term --cov=iaas_jwt -vv"
post = "coverage html"
uvicorn = "uvicorn  iaas_jwt.main:app --reload"